requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
tenacity>=8.2.0
openai>=1.0.0
google-genai>=0.1.0
anthropic>=0.5.0
ruff>=0.1.0
black>=23.0.0
